dsa: eager sparse-core + indexer kernel ops, pinned vs autograd (M-H1b)
Five registry ops (eager v1, per-sequence row-chunked, fp32-internal,
sync-free): dsa_index_scores (ReLU-weighted head-sum, causal+cross-seq
masking), dsa_topk (stable-desc sort — smallest-index ties, pad-safe),
dsa_sparse_attn_fwd (mask-form masked softmax over the padded-v MLA
tensors; emits the MASKED lse — the bwd + indexer-target anchor),
dsa_sparse_attn_bwd (probs from saved lse; deterministic dense
per-sequence GEMMs, no atomics), dsa_index_bwd (dI -> dq/dk/dwts with
ReLU sign recomputed, never materialized at (t,s,H)).

Pins now 7/7: scores vs reference 5.5e-8; topk exact; sparse fwd/bwd vs
autograd at bf16 tolerances incl. dv padding-column zeroness (with the
correctly-constructed zero-pad upstream gradient) and determinism-twice;
index bwd vs autograd (dw 1e-4). Optimized impls (triton index, triton
gather-absorbed core, sm90-gated FlashMLA seam, tiny-scale stock-flash
cross-validator) land in M-H2 behind the same op names.

+++ src/dataflow/tasks/dsa_reference.py
@@ -128,7 +128,7 @@ def dsa_sparse_attention_reference(
         q4 = q_full[lo:hi].view(1, L, h, qk).transpose(1, 2)
         k4 = k_full[lo:hi].view(1, L, h, qk).transpose(1, 2)
         v4 = v_pad[lo:hi].view(1, L, h, qk).transpose(1, 2)
-        m = add_mask[lo:hi, lo:hi].view(1, 1, L, L)
+        m = add_mask[lo:hi, lo:hi].view(1, 1, L, L).to(q4.dtype)
         o = F.scaled_dot_product_attention(q4, k4, v4, attn_mask=m)
         outs.append(o.transpose(1, 2).reshape(L, h * qk))
         lo = hi
